Rename analyse.js to beat-detector.js
Chrome blocks JS files named analyse.js lol I can't debug it when it's named that
