+++ src/beat-detector.js
